requirements.txt
fastapi==0.128.3
uvicorn==0.40.0
pandas==2.3.3
numpy==2.4.3
faiss-cpu==1.13.2
scikit-learn==1.8.0